+++ patzilla-ui/access/generic.js
@@ -1,4 +1,0 @@
-// -*- coding: utf-8 -*-
-// (c) 2013-2017 Andreas Motl, Elmyra UG
-//require('./10-ops-base.js');
-